separate cloudstore module
